fix: moved settings into new file
