+++ src/core/server/app/middleware/passport/oidc.ts
@@ -8,7 +8,8 @@ import { Strategy } from "passport-strategy";
 import { validate } from "talk-server/app/request/body";
 import { reconstructURL } from "talk-server/app/url";
 import { GQLUSER_ROLE } from "talk-server/graph/tenant/schema/__generated__/types";
-import { OIDCAuthIntegration, Tenant } from "talk-server/models/tenant";
+import { OIDCAuthIntegration } from "talk-server/models/settings";
+import { Tenant } from "talk-server/models/tenant";
 import { OIDCProfile, retrieveUserWithProfile } from "talk-server/models/user";
 import { upsert } from "talk-server/services/users";
 import { Request } from "talk-server/types/express";
