+++ src/core/server/graph/tenant/resolvers/auth_settings.ts
@@ -1,5 +1,5 @@
 import { GQLAuthSettingsTypeResolver } from "talk-server/graph/tenant/schema/__generated__/types";
-import { Auth } from "talk-server/models/tenant";
+import { Auth } from "talk-server/models/settings";
 
 const AuthSettings: GQLAuthSettingsTypeResolver<Auth> = {
   integrations: auth => auth.integrations,
