+++ src/core/server/graph/tenant/resolvers/facebook_auth_integration.ts
@@ -1,5 +1,5 @@
 import { GQLFacebookAuthIntegrationTypeResolver } from "talk-server/graph/tenant/schema/__generated__/types";
-import { FacebookAuthIntegration } from "talk-server/models/tenant";
+import { FacebookAuthIntegration } from "talk-server/models/settings";
 
 const FacebookAuthIntegration: GQLFacebookAuthIntegrationTypeResolver<
   FacebookAuthIntegration
