+++ src/core/server/graph/tenant/resolvers/google_auth_integration.ts
@@ -1,5 +1,5 @@
 import { GQLGoogleAuthIntegrationTypeResolver } from "talk-server/graph/tenant/schema/__generated__/types";
-import { GoogleAuthIntegration } from "talk-server/models/tenant";
+import { GoogleAuthIntegration } from "talk-server/models/settings";
 
 const GoogleAuthIntegration: GQLGoogleAuthIntegrationTypeResolver<
   GoogleAuthIntegration
