+++ src/core/server/graph/tenant/resolvers/local_auth_integration.ts
@@ -1,5 +1,5 @@
 import { GQLLocalAuthIntegrationTypeResolver } from "talk-server/graph/tenant/schema/__generated__/types";
-import { LocalAuthIntegration } from "talk-server/models/tenant";
+import { LocalAuthIntegration } from "talk-server/models/settings";
 
 const LocalAuthIntegration: GQLLocalAuthIntegrationTypeResolver<
   LocalAuthIntegration
