+++ src/core/server/graph/tenant/resolvers/oidc_auth_integration.ts
@@ -1,5 +1,5 @@
 import { GQLOIDCAuthIntegrationTypeResolver } from "talk-server/graph/tenant/schema/__generated__/types";
-import { OIDCAuthIntegration } from "talk-server/models/tenant";
+import { OIDCAuthIntegration } from "talk-server/models/settings";
 
 const OIDCAuthIntegration: GQLOIDCAuthIntegrationTypeResolver<
   OIDCAuthIntegration
