+++ src/core/server/graph/tenant/resolvers/sso_auth_integration.ts
@@ -1,5 +1,5 @@
 import { GQLSSOAuthIntegrationTypeResolver } from "talk-server/graph/tenant/schema/__generated__/types";
-import { SSOAuthIntegration } from "talk-server/models/tenant";
+import { SSOAuthIntegration } from "talk-server/models/settings";
 
 const SSOAuthIntegration: GQLSSOAuthIntegrationTypeResolver<
   SSOAuthIntegration
